requirements.txt
discord.py
requests
nest_asyncio
numpy
aiohttp
beautifulsoup4
matplotlib
python-dotenv
